Unifiying annotations
