+++ commercetools-internal-processors/src/main/resources/META-INF/services/javax.annotation.processing.Processor
@@ -11,5 +11,4 @@ io.sphere.sdk.annotations.processors.HasQueryModelImplementationAnnotationProces
 io.sphere.sdk.annotations.processors.HasBuilderAnnotationProcessor
 io.sphere.sdk.annotations.processors.HasUpdateActionsProcessor
 io.sphere.sdk.annotations.processors.HasUpdateActionProcessor
-io.sphere.sdk.annotations.processors.HasCustomUpdateActionProcessor
 io.sphere.sdk.annotations.processors.HasCustomUpdateActionsProcessor
