+++ commercetools-internal-processors/src/test/java/io/sphere/sdk/annotations/processors/generators/examples/ExampleResourceWithCustomUpdateAction.java
@@ -3,13 +3,10 @@ package io.sphere.sdk.annotations.processors.generators.examples;
 
 import io.sphere.sdk.annotations.*;
 
-import java.util.List;
-import java.util.Map;
-
 public interface ExampleResourceWithCustomUpdateAction {
 
 
-    @HasCustomUpdateAction(name = "customUpdateAction", fields = {
+    @HasUpdateAction(value = "customUpdateAction", fields = {
             @PropertySpec(name = "name", fieldType = String.class),
             @PropertySpec(name = "quantity", fieldType = Long.class),
             @PropertySpec(name = "slug", fieldType = String.class),
