--- commercetools-sdk-base/src/main/java/io/sphere/sdk/annotations/HasCustomUpdateAction.java
@@ -1,34 +1,0 @@
-package io.sphere.sdk.annotations;
-
-
-import java.lang.annotation.*;
-
-@Repeatable(HasCustomUpdateActions.class)
-@Target(ElementType.METHOD)
-@Retention(RetentionPolicy.CLASS)
-public @interface HasCustomUpdateAction {
-
-    String name();
-
-    PropertySpec[] fields();
-
-    /**
-     * This adds a {code @include.example <i>exampleBaseClass</i>#<i>updateActionTestName</i>()} javadoc tag
-     * to all generated update actions.
-     * <p>
-     * Where {@code <i>updateActionTestName</i>} is derived by un-capitalizing the update action name
-     * (e.g. for a {@code SetKey} update action the <i>updateActionTestName</i> is {@code setKey}.
-     * <p>
-     * This follows our convention to have integration tests for all of our update actions.
-     *
-     * @return the example base class name
-     */
-
-    String exampleBaseClass() default "";
-    FactoryMethod[] factoryMethods() default {};
-    CopyFactoryMethod[] copyFactoryMethods() default {};
-    Class[] superInterfaces() default {};
-    boolean makeAbstract() default false;
-
-
-}
